// src/store/lesson/types.ts
export enum Types {
  SAVE_SECTION_OF_COURSE = 'SAVE_SECTION_OF_COURSE',
  RESET = 'RESET',
  SAVE_DATA_SCRIPT = 'SAVE_DATA_SCRIPT',
  SAVE_HIGHT_SCORE = 'SAVE_HIGHT_SCORE',
  SHOW_REPORT = 'SHOW_REPORT',
  SHOW_REPORTED = 'SHOW_REPORTED',
  SHOW_SUMARY = 'SHOW_SUMARY',
  ADD_CORRECT = 'ADD_CORRECT',
  ADD_WRONG = 'ADD_WRONG',
  RESET_CORRECT = 'RESET_CORRECT',
  SHOW_CHEER = 'SHOW_CHEER',
  SHOW_FIRST_CHEER_CORRECT = 'SHOW_FIRST_CHEER_CORRECT',
  SHOW_SECOND_CHEER_CORRECT = 'SHOW_SECOND_CHEER_CORRECT',
  SHOW_CHEER_WRONG = 'SHOW_CHEER_WRONG'
}